refactor(viewer): split filter evaluator + builder to reduce module size (#921)
Review follow-up on #931 (CodeRabbit size-cap, ~400 lines/module):
- Extract per-rule value resolution/matching out of filter-evaluate.ts
  into filter-match.ts (flatten/match/material/classification/elevation
  helpers). The evaluator keeps iteration + orchestration; __internal
  re-exports the helpers so the existing tests are unchanged.
- Extract the chip editors + RuleRow + op constants out of
  SearchModal.filter.builder.tsx into SearchModal.filter.editors.tsx.
  The builder (379 lines) now owns only toolbar / presets / run state.

No behaviour change; filter tests (62) and typecheck pass.
